requirements.txt
allure-python-commons
requests
pytest~=8.4.2
selenium==4.37.0
webdriver-manager==4.0.2
